api_yamdb/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
django-filter==21.1
django==2.2.16
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.0
mccabe==0.6.1
packaging==21.2; python_version >= '3.6'
pluggy==0.13.1
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
pyjwt==2.3.0
pyparsing==2.4.7
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
pytz==2021.3
requests==2.26.0
semantic-version==2.8.5
setuptools-rust==0.12.1
setuptools==58.5.3; python_version >= '3.6'
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
gunicorn==20.0.4
psycopg2-binary==2.8.6